Renamed player modules to character
